try to fix calc.py

+++ calc.py
@@ -2,15 +2,14 @@
 В предыдущей задаче от вас требовалось реализовать функцию convert(expr),
 которая принимала в качестве единственного аргумента математическое выражение
 и возвращала его запись в ОПН.
-
 Теперь от вас требуется написать программу "Калькулятор".
-
 Он должен работать в двух режимах:
     в режиме Python-модуля
     в режиме исполняемого Python-скрипта.
 '''
+
 from task_rpn import convert
-from operator import add, sub, mul, truediv
+from operator import add, sub, mul, truediv, pow
 
 
 OP_DICT = {'^': pow, '*': mul, '/': truediv, '+': add, '-': sub}
@@ -25,7 +24,7 @@ def calc(expr):
             b, a = stack.pop(), stack.pop()
             stack.append(OP_DICT[char](a, b))
         else:
-            stack.append(float(char))
+            stack.append(int(char))
 
     return sum(stack)
